WIP: version 0.0.7 release
