+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "wemake-python-styleguide"
-version = "0.0.6"
+version = "0.0.7"
 description = "Opinionated styleguide that we use in wemake.services"
 
 license = "MIT"
