+++ wemake_python_styleguide/visitors/complexity/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+# TODO: implement TooDeepNestingViolation, TooManyBranchesViolation
